refactor 예약 내역 리팩토링 #5 - Feign Client로 받아오는 값들 Dto로 변경(Entity -> Dto)
