--- com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/domain/model/Performance.java
@@ -1,10 +1,0 @@
-package com.culture_ticket.client.reservation_payment.domain.model;
-
-import lombok.Getter;
-
-@Getter
-public class Performance {
-    String title;
-    String venue;
-    String performanceStatus;
-}
