--- com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/domain/model/User.java
@@ -1,11 +1,0 @@
-package com.culture_ticket.client.reservation_payment.domain.model;
-
-import lombok.Getter;
-
-@Getter
-public class User {
-
-    String username;
-    String nickname;
-    String phone;
-}
