+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/infrastructure/client/PerformanceClient.java
@@ -1,9 +1,9 @@
 package com.culture_ticket.client.reservation_payment.infrastructure.client;
 
 import com.culture_ticket.client.reservation_payment.common.ResponseDataDto;
-import com.culture_ticket.client.reservation_payment.domain.model.Performance;
-import com.culture_ticket.client.reservation_payment.domain.model.Seat;
-import com.culture_ticket.client.reservation_payment.domain.model.TimeTable;
+import com.culture_ticket.client.reservation_payment.infrastructure.dto.PerformanceResponseDto;
+import com.culture_ticket.client.reservation_payment.infrastructure.dto.SeatResponseDto;
+import com.culture_ticket.client.reservation_payment.infrastructure.dto.TimeTableResponseDto;
 import java.util.UUID;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,11 +13,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface PerformanceClient {
 
     @GetMapping("/performances/{performanceId}")
-    ResponseDataDto<Performance> getPerfomance(@PathVariable UUID perfomanceId);
+    ResponseDataDto<PerformanceResponseDto> getPerfomance(@PathVariable UUID perfomanceId);
 
     @GetMapping("/seats/{seatId}")
-    ResponseDataDto<Seat> getSeat(@PathVariable UUID seatId);
+    ResponseDataDto<SeatResponseDto> getSeat(@PathVariable UUID seatId);
 
     @GetMapping("/timetables/{timeTableId}")
-    ResponseDataDto<TimeTable> getTimeTable(@PathVariable UUID timeTableId);
+    ResponseDataDto<TimeTableResponseDto> getTimeTable(@PathVariable UUID timeTableId);
 }
