+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/infrastructure/client/UserClient.java
@@ -1,7 +1,7 @@
 package com.culture_ticket.client.reservation_payment.infrastructure.client;
 
 import com.culture_ticket.client.reservation_payment.common.ResponseDataDto;
-import com.culture_ticket.client.reservation_payment.domain.model.User;
+import com.culture_ticket.client.reservation_payment.infrastructure.dto.UserResponseDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,5 +10,5 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface UserClient {
 
     @GetMapping("/info/{userId}")
-    ResponseDataDto<User> getUser(@PathVariable long userId);
+    ResponseDataDto<UserResponseDto> getUser(@PathVariable long userId);
 }
