+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/infrastructure/dto/PerformanceResponseDto.java
@@ -1,0 +1,10 @@
+package com.culture_ticket.client.reservation_payment.infrastructure.dto;
+
+import lombok.Getter;
+
+@Getter
+public class PerformanceResponseDto {
+    String title;
+    String venue;
+    String performanceStatus;
+}
