+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/infrastructure/dto/SeatResponseDto.java
@@ -1,25 +1,25 @@
-package com.culture_ticket.client.reservation_payment.domain.model;
+package com.culture_ticket.client.reservation_payment.infrastructure.dto;
 
 import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-public class Seat {
+public class SeatResponseDto {
     UUID timeTableId;
     String seatClass;
     Integer seatNumber;
     long price;
 
     @Builder
-    private Seat(UUID timeTableId, String seatClass, Integer seatNumber, long price) {
+    private SeatResponseDto(UUID timeTableId, String seatClass, Integer seatNumber, long price) {
         this.timeTableId = timeTableId;
         this.seatNumber = seatNumber;
         this.seatClass = seatClass;
         this.price = price;
     }
 
-    public static Seat from(Seat seat) {
+    public static SeatResponseDto from(SeatResponseDto seat) {
         return builder()
             .timeTableId(seat.getTimeTableId())
             .seatClass(seat.getSeatClass())
