+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/infrastructure/dto/UserResponseDto.java
@@ -1,0 +1,11 @@
+package com.culture_ticket.client.reservation_payment.infrastructure.dto;
+
+import lombok.Getter;
+
+@Getter
+public class UserResponseDto {
+
+    String username;
+    String nickname;
+    String phone;
+}
