+++ com.culture-ticket.client.reservation-payment/src/main/resources/application.yml
@@ -24,8 +24,3 @@ eureka:
   client:
     service-url:
       defaultZone: http://localhost:19090/eureka/
-
-logging:
-  level:
-    org:
-      springframework: DEBUG
